+++ projects/rocprofiler-compute/src/argparser.py
@@ -588,16 +588,18 @@ Examples:
         required=False,
         metavar="",
         dest="pc_sampling_interval",
-        default=1048576,
+        default=None,
+        type=int,
         base_action="store",
         action=ExperimentalAction,
         experimental_enabled=experimental_enabled,
         feature_label="PC Sampling",
         help=(
             "\t\t\tSet the interval of pc sampling.\n"
-            "\t\t\t  For stochastic sampling, the interval is in cycles.\n"
-            "\t\t\t  For host_trap sampling, the interval is in microsecond "
-            "(DEFAULT: 1048576)."
+            "\t\t\t  For stochastic sampling, the interval is in cycles; it "
+            "must be a power of 2 and at least 65536 (DEFAULT: 1048576).\n"
+            "\t\t\t  For host_trap sampling, the interval is in microseconds "
+            "(DEFAULT: 512)."
         ),
     )