+++ projects/rocprofiler-compute/tests/test_pc_sampling.py
@@ -94,8 +94,6 @@ def test_pc_sampling_stochastic(binary_handler_profile_rocprof_compute, monkeypa
         "21",
         "--pc-sampling-method",
         "stochastic",
-        "--pc-sampling-interval",
-        "1048576",
     ]
 
     workload_dir = common.get_output_dir()
@@ -141,8 +139,6 @@ def test_multi_rank_pc_sampling_only(
         "21",
         "--pc-sampling-method",
         "host_trap",
-        "--pc-sampling-interval",
-        "256",
     ]
 
     _, stdout, stderr = binary_handler_profile_rocprof_compute(
@@ -231,8 +227,6 @@ def test_pc_sampling_profile_then_analyze(
         "21",
         "--pc-sampling-method",
         "host_trap",
-        "--pc-sampling-interval",
-        "256",
     ]
 
     workload_dir = common.get_output_dir()